<==== Mercurial data ====>
Branch: feature_3948_webclient

+++ common/src/plugins/resource/avi/thumbnails_archive_delegate.cpp
@@ -79,14 +79,14 @@ QnAbstractMediaDataPtr QnThumbnailsArchiveDelegate::getNextData()
     if (!delegateForMediaStep) 
     {
         qint64 startTime = m_baseDelegate->startTime();
-        if (startTime != AV_NOPTS_VALUE) {
+        if (startTime != qint64(AV_NOPTS_VALUE)) {
             qint64 outerSteps = (startTime - m_currentPos) / m_frameStep;
             if (outerSteps > 0)
                 m_currentPos += outerSteps * m_frameStep;
         }
 
         qint64 seekRez = m_baseDelegate->seek(m_currentPos, true);
-        while (seekRez == -1 && m_currentPos < m_rangeEnd && m_rangeEnd != AV_NOPTS_VALUE) {
+        while (seekRez == -1 && m_currentPos < m_rangeEnd && m_rangeEnd != qint64(AV_NOPTS_VALUE)) {
             m_currentPos += m_frameStep;
             seekRez = m_baseDelegate->seek(m_currentPos, true);
         }
